// types/generated/graphql.tsx
export type GetDateFormulasQuery = { __typename?: 'query_root', date_calculation_dateformula: Array<{ __typename?: 'date_calculation_dateformula', dateId?: string | null, dateName?: string | null, adjustmentDay?: number | null, adjustmentMonth?: number | null, adjustmentYear?: number | null }> };